moved fermion spin generators under hamiltonians
The only place to move the spin-generators is under the hamiltonians
package.  If moved to utils there is a cyclic import because `up_index`
and `down_index` will be used by `hamiltonians._hubabrd.py` and
`hamiltonians._mean_field_dwave.py`.

Maybe the `hamiltonians` package is a batter place for it since these
spin operators are technically fermionic Hamiltonians correspond to an
observable one likely wants to measure.