RDM mapping functions with extensive testing
This PR provides functionality to convert marginals in various forms to
one another.  All inputs attempt to obey the tensor ordering convention
used throughout openfermion.

In order to have quality tests we've also added 2-hole-marginals,
particle-hole-marginals, and the code that generated them.

I am completely open to moving the functions somewhere else in the
package if everyone thinks the `measurements` submodule is an
inappropriate place for the methods.