Add README link to run Example Notebooks in MyBinder (#496)
* Add link to run Example Notebooks in MyBinder

The MyBinder website provides free, cloud-based hosting of jupyter notebooks, built from public git repos!! :)

I was able to just enter this repo's URL into the builder on https://mybinder.org/, and it correctly set up the repo based on the Dockerfile and everything seems to work! :) This means with one-click, people can open a browser-based environment and start interacting with the tutorial! :)

* Oops, fix typo in README.rst