Hartree-fock (#583)
* Upload of diazene data

Diazene input data for HFVQE.

nickrubin@google.com

* Gradient based Resetricted Hartree-Fock

Infrastructure for implementing various kinds of Hartree-Fock theories.

Currently we support restricted-Hartree-Fock.  WIP is unrestricted-, and
generalized-Hartree-Fock theories.

This provides the user with a way to converge Hartree-Fock without having
to know many of the details and tricks for getting SCF procedures to
converge.

The downside is that this procedure is slow.  We evaluate everything in
the spin-orbital basis for generality and so we can reuse gradient code
for UHF and GHF.

* Increasing coverage and definite imports

* Making my code ugly to satisfy check/format

Okay...some things got better...