Implement basics for representability library
This is the foundation of constructing the SDP projections found in
ArXiv:1801.03524.

We implement a way to index into tensors where the index is some composite
index or map. An example of this would be the supermatrix representation
of the 2-RDM as a matrix.  Each index of this matrix is a geminal index.

We also provide a way to abstractly represent operations over vectorized
tensor spaces called a DualBasis object.  We can use this to construct
the SDPs associated with 2-RDM projection techniques and variational
2-RDM optimizations.