Dual basis (#588)
* Upload of diazene data

Diazene input data for HFVQE.

nickrubin@google.com

* Implement basics for representability library

This is the foundation of constructing the SDP projections found in
ArXiv:1801.03524.

We implement a way to index into tensors where the index is some composite
index or map. An example of this would be the supermatrix representation
of the 2-RDM as a matrix.  Each index of this matrix is a geminal index.

We also provide a way to abstractly represent operations over vectorized
tensor spaces called a DualBasis object.  We can use this to construct
the SDPs associated with 2-RDM projection techniques and variational
2-RDM optimizations.

* remove tqdm tracking

* uncaught lint

The results of running pylint locally vs on CI seems different.
Everything can pass on my local machine but then it hits the cloud and
clearly it misses stuff...weird...

* remove comment

* add typing to arg and returns

* remove object inheritence

* Formating for new pylinrc file

looks like pylintrc was changed in the most recent commit on master

Co-authored-by: Ryan Babbush <ryanbabbush@gmail.com>