change to numpy  for linalg and remove message
The NotImplementedError seems to be an issue.  The correct error is
raised when run in isolation but in production it seems to fail the
test.

The segfaulting lih test is changed to numpy for stability and instead
of converting the hamiltonian to a Matrix object we change it to an
ndarray.