Reorganization of OpenFermion (#620)
* Reorganization of OpenFermion

* New module structure
* No functionality changes

* Adding RDM testing files that were missed

These files were missed because of the git ignore.  Needed to force add.

* format check correction

* Remove cyclic import so test pass on 3.6

Interestingly in the move to python 3.7 they patched some cyclic
import issues.

* Increasing coverage

Up to date coverage.  We have ~50 lines uncovered.

* format and linting

* Added coverage to circuits

* Linting and formatting

* More linting and formatting

* Added coverage to ops

* F or mat

* Fixed low-rank coverage issue, fixed bug in SymbolicOperator

* F o r m at

* Form at

* One last format

* Improved coverage in measurements and linalg

* For mat

* Made import explicit, seems to pass

* Removed import *

* Fixed shorthand numpy

* Found another shorthand numpy

* Fixing linter

* Formatting and linter

* Fixing coverage of testing

* Formatting and new test coverage

Coverage reduces untouched code.

Issues for the pragma no covers will be addressed.

One pragma no cover is because OpenFermion can't reach the
exception by design.

* Removed unnecessary class features in testing tests

* Final line coverage and linting

* format that wasn't caught

sometimes the linter seems to ignore things?

* Reduce resource requirements for the sparse test

I also change to pytest for testing the NotImplementedError
exception on the get_molecular_data routine.

* change to numpy  for linalg and remove message

The NotImplementedError seems to be an issue.  The correct error is
raised when run in isolation but in production it seems to fail the
test.

The segfaulting lih test is changed to numpy for stability and instead
of converting the hamiltonian to a Matrix object we change it to an
ndarray.

Co-authored-by: Tom O'Brien <teobrien@google.com>