Update .pylintrc for Pylint 3.x
The Pylint config file for 3.x has considerably changed from 2.x.

The `.pylintrc` in this commit has many checks turned off. I ran
`pylint` 3.3.3 repeatedly on the codebase, checking whether the
warnings looked serious. For many that looked like probable code
errors, I opened issues, but the work is incomplete. In the end I
started disabling most checks because there was too much to do and we
were aiming to do a release asap.

We need to go back, re-enable checks in .pylintrc one by one, and
correct the warnings in the code.