Minor enhancements to README file & fix to setup.py (#934)
* Minor enhancements to README file

This fixes a couple of typos, adds a couple of badges, and tries to
adjust some spacing issues. (The result looks better on PyPI than on
GitHub because of GitHub's poor support for reStructuredText.)

* Fix truncation of readme file contents

The original code ended up cutting off the first line of the
readme.rst file. Replaced it with something that doesn't do that and
is a bit simpler.