Change this CodeQL config to a nightly run only
I can't seem to get rid of or disable the default CodeQL configuration in GitHub, and so we ended up with 2 CodeQL runs every time something happened. After spending > 2 hours trying to figure out a way to disable it, I'm giving up. Instead, I'm going to make this a nightly run of CodeQL, so that it's run even on days when nothing else changes in the repository on the off chance that CodeQL itself is updated.