Make a few minor improvements to the stats-saving workflow (#983)
* Use var for data branch name & create if needed

When installing this workflow in a new repo, the data branch would
have to be created manually -- something that's easy to forget. This
now creates it automatically if it doesn't exist. (Not needed for
OpenFermion because it's already here, but for others it may be.)

This also now prints a note in the job summary to point to the report.