Add OSV code scanner and make minor adjustments to CI (#992)
* Run all nightly tests from a common top-level workflow

This removes the individual cron schedule entries from the workflows
that run at night, and instead invokes them from `nightly.yaml`.

* Rename workflows to use .yaml extension consistently

`.yaml` is actually the official file extension for YAML.

* Add OSV scanner workflow

* Adjust names and trigger conditions

* Tell yamllint the long lines are okay

* Rework organization some more

* Set up another problem matcher

* Add .editorconfig file (#990)

This adds settings based on some common-sense values and the
OpenFermion project's current conventions, such as line length and use
of spaces instead of tabs.

* Add shellcheck for shell script linting (#991)

* Add .shellcheckrc

This adds a config file for .shellcheckrc, and sets a few optional
flags to try to make shellcheck slightly more thorough.

* Add shellcheck problem matcher

* Add a heading to help break up the text a little

* Add shellcheck to CI checks

* Edit CONTRIBUTING.md (#993)

This adds a warning about the CLA implications for GenAI code, and
tries to add some more info about the contribution process.

* Add markdownlint config file (#995)

This adds a `.markdownlintrc` configuration file for
[markdownlint](https://github.com/igorshubovych/markdownlint-cli), a
popular linter for Markdown files. The settings here try to be as
close to the Google style as possible while still being compatible
with GitHub-flavored Markdown.

This same markdownlint file is used in Cirq and other Quantumlib
projects.

* Fix warnings reported by shellcheck (#997)

There are no functional changes in this commit; all changes are to
address lint reported by shellcheck.

* CI checks need to be run if ci.yml changes (#998)

The tests for changed files needs to consider changes to the workflow
definitions too, because a change to the workflow definition could
involve changing options or other aspects of the CI tests that might
change the check results.

* Add problem matchers for Mypy, Pytest, and Black (#1000)

* Add problem matchers for mypy and pytest

* Add problem matcher for Black

* Pass input sha to format-incremental

If the user supplied a sha in a manual workflow run, we need to pass
that to `format-incremental` or else it won't compare files to the
correct revision.

* Prefer single quotes in YAML

Due to how GitHub Actions syntax works, I'm finding it safer to use
single quotes for YAML values and (usually) double quotes in shell
scripts.

* Tighten format of "run:" for one-liners

It seems slightly reasier to read the code whenn single-line shell
commands are put on the same line as the `run:`, instead of being put
on a separate line.

* More setup, editing, configuration, renaming, tweaking