Run OSV scan on PRs if requirements files are changed (#1004)
* Run OSV scan on PRs if requirements files are changed

The sample workflows for the OSV scanner make it run on every PR, but
that doesn't make sense to me; better would be to run it only if a
requirements file is changed in a PR. So, that's what this does.

* Remove --skip-git option from osv-scanner

The documentation is poor, and the meaning of `--skip-git` is
ambiguous, and it doesn't seem to reduce performance noticeably, so
lets' just skip that option.

* Rename a couple of jobs for consistency with others