Ignore false positives from osv-scanner
OSV prior to version 2.0 is unable to parse pip version specs
correctly, as explained by one of the developers in the following
comment from January 2025:

https://github.com/google/osv-scanner/issues/1483#issuecomment-2585999293

The suggested workaround is to configure osv-scanner to ignore the particular
cases it complains about. The following are all about NumPy, because
osv-scanner can't understand the version spec "numpy>=1.24,<2.0" and
therefore raises errors about all versions of NumPy, including very old ones.
Ignoring these specific dependencies is okay because we will never use the
old versions of NumPy and it doesn't block detection of future new
vulnerabilities.