Trigger on auto_merge_enabled
The trimming needs to be performed before the PR is merged so that the commit
message is based on the trimmed version. This prevents us from using events
such as the PR getting closed or merged. Triggering on merge queue events is
also problematic, because a merge_queue event doesn't have the equivalent of
"pull_request.user" and we need that to test if the PR came from Dependabot.
So instead, this workflow triggers when auto-merge is enabled for the PR.