Match data types when creating array of zeros
In function `get_tensors_from_integrals`, the call to `numpy.zeros`
didn't supply data type information. This resulted in the type
defaulting to a float type, which then led to warnings about losing
precision in the multiplications later in the method. The solution is to
match the types of the values in the call to `zeros` to the types of
other arrays.