Remove CodeQL scanner & move others to weekly.yaml
GitHub enhanced their CodeQL scanner and its configuration options since
the time this was written, and it's now easier to simply use the
repository settings for setting up CodeQL. No need to maintain our own
workflow anymore.

Also, the other scanners (OSV and Scorecard) are run on every PR, so
there is not much value in running it nightly too. I moved them to a new
weekly workflow because it's still worth running them regularly.