Make CI change include changes to tools & check scripts
If a script in `./check/` or `./dev_tools/` is changed, it may affect
the results of running CI tests. To account for that, this PR expands
the file change filter to detect any changes in those directories.

In addition, this widens the files tested in `.github/workflows` to
include any file changes.

The result is that any change in the dev_tools, scripts, or the CI
workflows will trigger all the tests, not only the file-type-specific
tests.