Tell actionlint that image ubuntu-24.04-arm is valid
The addition of testing on ubuntu-24.04-arm causes `actionlint` to
produce a false positive error that the runner name is invalid. It
happens because that particular runner image is one of the partner
images from https://github.com/actions/partner-runner-images

This `actionlint` configuration file defines the runner name as valid.