Update hadolint configuration (#1142)
The `hadolint` warning DL3009 is only relevant when optimizing a docker
container's size. It's not needed for security or other reasons. Not
only is this rule annoying; following the rule can also increase docker
image build times. The benefit does not seem to outweight the drawbacks,
so I added it to the list of ignored rules.

Conversely, DL3047 (about the syntax of CMD statements) is probably
better to keep, as a matter of modern practices surrounding Dockerfiles.