Fix #1120: change unsafe regex in symbolic_operator.py (#1140)
CodeQL scan [report
#566](https://github.com/quantumlib/OpenFermion/security/code-scanning/566)
flagged a regex used on line 165 of
`src/openfermion/ops/operators/symbolic_operator.py` as being potential
subject to a DoS attach. The warning is this:

```python
pattern = r'(.*?)\[(.*?)\]'  # regex for a term
for match in re.findall(pattern, long_string, flags=re.DOTALL):
```

```
This regular expression that depends on a user-provided value
 may run slow on strings with many repetitions of 'a'.
This regular expression that depends on a user-provided value
 may run slow on strings starting with '[' and with many repetitions of '[a'.
This regular expression that depends on a user-provided value
 may run slow on strings with many repetitions of 'a'.
This regular expression that depends on a user-provided value
 may run slow on strings starting with '[' and with many repetitions of '[a'.
```

This changes the regular expression to avoid `.*` yet still be able to
match the same patterns as before. Additional tests in
`symbolic_operator_test.py` verify that this will parse strings
correctly.