Tell actionlint that image ubuntu-24.04-arm is valid (#1136)
The addition of testing on ubuntu-24.04-arm causes `actionlint` to
produce a false positive error about the runner name being invalid. The
error happens because that particular runner image is one of the partner
images from https://github.com/actions/partner-runner-images and
apparently `actionlint` does not know the names of those by default.

This `actionlint` configuration file defines the runner name as valid.