Fix #1137 and #908: replace README.rst with Markdown version & update content (#1143)
This replaces the reStructuredText version of the README file with a
Markdown version, for better compatibility with our practices in our
other repos. In addition, it removes the OpenFermion logo to follow the
updated branding guidelines for Quantum AI. It also slightly reorganizes
the content, as well as provides a BibTeX version of the citation info,
and finally, adds the standard Google disclaimer and our QAI logo at the
bottom.